+++ attributes/default.rb
@@ -1,3 +1,3 @@
 
-default['chef-zki']['zookeeper']['server'] = '127.0.0.1:2181'
+default['zookeeper-bridge']['zookeeper']['server'] = '127.0.0.1:2181'