+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'chef-zki'
+name             'zookeeper-bridge'
 maintainer       'Onddo Labs, Sl.'
 maintainer_email 'team@onddo.com'
 license          'Apache 2.0'
-description      'Installs/Configures chef-zki'
+description      'Installs/Configures zookeeper-bridge'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'