+++ test/kitchen/cookbooks/zookeeper-bridge_test/metadata.rb
@@ -1,11 +1,11 @@
-name             'chef-zki_test'
+name             'zookeeper-bridge_test'
 maintainer       'Onddo Labs, Sl.'
 maintainer_email 'team@onddo.com'
 license          'Apache 2.0'
-description      'Installs/Configures chef-zki_test'
+description      'Installs/Configures zookeeper-bridge_test'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
-depends 'chef-zki'
+depends 'zookeeper-bridge'
 
 depends 'zookeeper'