+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-chef-zki CHANGELOG
+zookeeper-bridge CHANGELOG
 ==================
 
-This file is used to list changes made in each version of the chef-zki cookbook.
+This file is used to list changes made in each version of the zookeeper-bridge cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of chef-zki
+- [your_name] - Initial release of zookeeper-bridge
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.